+++ AlgoMonsterDaily/BoilerPlate.cs
@@ -1,0 +1,8 @@
+﻿public static class BoilerPlate
+{
+    public static void Main()
+    { 
+        
+    }
+}
+
