--- AlgoMonsterDaily/DFS.cs
@@ -1,8 +1,0 @@
-﻿namespace Algorithms.DFS
-{
-    // Search algorithm
-    public class DFS
-    {
-        
-    }
-}
