+++ AlgoMonsterDaily/LinkedNodes.cs
