--- AlgoTests/LeetCode_Test.cs
@@ -1,10 +1,0 @@
-using System.Data;
-
-namespace AlgoTests
-{
-    [TestClass]
-    public class LeetCode_Test
-    {
-       
-    }
-}
