+++ AlgoTests/OtherTestSites.cs
@@ -1,10 +1,51 @@
 using DataStructures;
-using System.Data;
-
-namespace AlgoTests
+using TestHelpers;
+namespace GeeksForGeeks.Tests
 {
     [TestClass]
-    public class LRUCache_Leetcode_Test
+    public class GFGPractice
+    {
+        [TestMethod]
+        [DataRow("0 -1 2 -3 1")]
+        public void FindTripletsWithZeroSum(string data)
+        {
+            var arr = TestHelper.SplitWords(data).Select(int.Parse).ToArray();
+
+            var triplets = Triplets.FindTriplets(arr, arr.Length);
+
+            if (triplets.Count == 0) { Console.WriteLine("none exists"); }
+        }
+    }    
+}
+namespace Algorithms.Tests
+{
+    [TestClass]
+    public class NeetCodeAdvancedCourse
+    {
+        #region Kadane's Algorithm
+        [TestMethod]
+        [DataRow("4 -1 2 -7 3 4", 7)]
+        [DataRow("-8 6 17 -4 9 -2 -15 1 -1 7", 28)]
+        public void FindNonEmptySubarrayWithLargestSum(string data, int expected)
+        {
+            var arr = TestHelper.SplitWords(data).Select(int.Parse).ToArray();
+
+            var kadaneWindow = Kadane.KadaneWindow(arr);
+            foreach (var item in kadaneWindow)
+            {
+                Console.WriteLine(item);
+            }
+
+            Assert.AreEqual(Kadane.KadaneBruteForce(arr), expected);
+            Assert.AreEqual(Kadane.KadaneLinear(arr), expected);
+        }
+        #endregion
+    }
+}
+namespace Leetcode.Tests
+{
+    [TestClass]
+    public class LRUCacheTest
     {
         [TestMethod]
         [DataRow(2)]
@@ -69,8 +110,8 @@ namespace AlgoTests
                 Assert.AreEqual(cache.Get(i), i);
                 outlier = i++;
             }
-            
-            cache.Put(outlier, outlier+7);
+
+            cache.Put(outlier, outlier + 7);
             Assert.AreEqual(cache.Get(2), -1);
             cache.Put(4, 8);
             Assert.AreEqual(cache.Get(4), 8);
